+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-05-21
+### Changed
+* Require email verification by default and add "require_verification" config
+
+### Added
+* Add `!json_decode` tag
+* Add optional userinfo logging
+* Add OpenID Connect session custom HTTP header in config
+
+### Fixed
+* Catch "uncaughtException" after served
+* Set requestTimeout: 0 to HTTP/1.1 server option for supporting Node.js 18
+
 ## [0.3.1] - 2023-05-20
 ### Fixed
 * Use Node.js 18 in vercel/pkg build for ARM
@@ -45,7 +58,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Support TLS certificate hot reload
 * Add "migrate-config" subcommand
 
-[Unreleased]: https://github.com/nwtgck/rich-piping-server/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/nwtgck/rich-piping-server/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/nwtgck/rich-piping-server/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/nwtgck/rich-piping-server/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/nwtgck/rich-piping-server/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/nwtgck/rich-piping-server/compare/v0.2.1...v0.2.2